requirements.txt
discord.py==2.6.3
uvicorn==0.35.0
fastapi==0.116.1
tzdata==2025.2